# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'ingot-forms API'
  description: 'Forms management backend. A form is a single fillable document: one immutable definition plus one data set with a required expire date. Lifecycle: empty → draft (repeatable lenient saves) → confirmed (strict validation, locked forever). Expired forms answer 410 everywhere. Every error response is an RFC 9457 `application/problem+json` document.'
  license:
    name: MIT
    identifier: MIT
  version: 0.2.0
servers:
  -
    url: /
paths:
  '/api/manage/form-templates/{template}/current':
    put:
      summary: 'Put a pair of versions in use'
      description: 'The pair is stated whole: naming a definition and no presentation means this template shows nothing from here on, not that it keeps the presentation it had. Going back to an earlier pair is this same call pointing the other way, and costs no new version.'
      operationId: activateTemplateVersions
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateVersionsRequest'
      responses:
        '204':
          description: 'In use.'
        '404':
          $ref: '#/components/responses/TemplateVersionNotFound'
        '422':
          description: 'The pair does not fit: the findings point at the item the presentation shows and the definition does not declare.'
...
  /api/manage/form-templates:
    get:
      summary: 'List the form templates'
      description: 'Newest first. Each entry names the pair of versions new forms made from it are given — as numbers, which is what the two history addresses take.'
      operationId: listFormTemplates
      responses:
        '200':
          description: 'Every template in the catalogue.'
          content:
            application/json:
              schema:
                properties:
                  templates: { type: array, items: { properties: { id: { type: string, format: uuid }, name: { type: string }, createdAt: { type: string, format: date-time }, createdBy: { type: [string, 'null'] }, definition: { type: integer }, presentation: { type: [integer, 'null'] } }, type: object } }
                type: object
    post:
      summary: 'Create a form template'
      description: 'A template is a name and the one pair of documents new forms made from it get. Both documents arrive here and become version 1 of their own history, already in use. Everything afterwards publishes into one history at a time and is put in use as its own deliberate act.'
      operationId: createFormTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFormTemplateRequest'
      responses:
        '201':
          description: 'Template created, holding version 1 of each document and using them.'
          headers:
            Location:
              description: 'Path of the created template.'
              schema:
                type: string
          content:
            application/json:
              schema:
                properties:
                  id: { type: string, format: uuid }
                  definition: { type: integer, example: 1 }
                  presentation: { type: [integer, 'null'], example: 1 }
                type: object
        '400':
          $ref: '#/components/responses/MalformedJson'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          description: 'The request, the definition, or the presentation beside it is not valid — including a presentation that does not fit the definition it came with.'
...
  '/api/manage/form-templates/{template}/definitions':
    get:
      summary: "List a template's definitions"
      description: 'Newest first, holding the numbers and how each got there — never the documents. Which of these is in use is `GET …/{template}`.'
      operationId: listTemplateDefinitions
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      responses:
        '200':
          description: 'The definition history.'
        '404':
          $ref: '#/components/responses/FormTemplateNotFound'
    post:
      summary: 'Publish a definition into a template'
      description: 'The answer is the number it was published as, which is the one thing the client could not know. Forms created now are still made of whatever `PUT …/current` last named.'
      operationId: publishTemplateDefinition
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishDefinitionRequest'
      responses:
        '201':
          description: 'Published. Nothing in use has changed.'
          content:
            application/json:
              schema:
                properties:
                  definition: { type: integer, example: 2 }
                type: object
        '404':
          $ref: '#/components/responses/FormTemplateNotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          description: 'The definition breaks the meta-schema or a semantic rule.'
  '/api/manage/form-templates/{template}/presentations':
    get:
      summary: "List a template's presentations"
      description: 'The other history, numbered apart from the definitions — a relabelled option is not a new model, and the two counts say so.'
      operationId: listTemplatePresentations
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      responses:
        '200':
          description: 'The presentation history.'
        '404':
          $ref: '#/components/responses/FormTemplateNotFound'
    post:
      summary: 'Publish a presentation into a template'
      description: 'The answer is the number it was published as. It is judged against the definition currently in use — a presentation showing an item that definition does not declare is refused here, with the findings pointing at the item — and nothing about what is in use changes.'
      operationId: publishTemplatePresentation
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishPresentationRequest'
      responses:
        '201':
          description: 'Published. Nothing in use has changed.'
          content:
            application/json:
              schema:
                properties:
                  presentation: { type: integer, example: 4 }
                type: object
        '404':
          $ref: '#/components/responses/FormTemplateNotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          description: 'The presentation breaks its meta-schema, or does not fit the definition currently in use.'
...
  '/api/forms/{id}/presentation':
    get:
      summary: 'Read how the form is shown'
      description: "The document as it was set. Codes are served unresolved: which language a client shows is the client's to decide, so nothing here reads Accept-Language."
      operationId: getFormPresentation
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      responses:
        '200':
          description: 'The presentation document.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormPresentation'
        '404':
          description: 'Unknown form, or a form nobody has said anything about yet.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                presentation-not-set:
                  summary: 'The form exists; how to show it was never said'
                  value: { type: 'urn:problem:ingot-forms:presentation-not-set', title: 'The form has no presentation.', status: 404 }
        '410':
          $ref: '#/components/responses/FormGone'
...
  '/api/manage/form-templates/{template}':
    get:
      summary: 'Read a form template'
      description: 'The pair in use is given as version numbers, which is what `PUT …/current` takes and what the two history addresses read. `forms` counts what is made of any version this template has ever published — the number a delete is refused over.'
      operationId: readFormTemplate
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      responses:
        '200':
          description: 'The template.'
          content:
            application/json:
              schema:
                properties:
                  id: { type: string, format: uuid }
                  name: { type: string }
                  createdAt: { type: string, format: date-time }
                  createdBy: { type: [string, 'null'] }
                  definition: { description: 'The definition version new forms are made of.', type: integer }
                  presentation: { description: 'The presentation version that shows them, or null when this template shows nothing.', type: [integer, 'null'] }
                  forms: { description: 'How many forms are made of any version this template ever published.', type: integer }
                type: object
        '404':
          $ref: '#/components/responses/FormTemplateNotFound'
...
components:
  schemas:
...
    FormPresentation:
      description: 'How a form is shown. Its authoritative contract is the meta-schema served at `GET /api/schemas/presentation`: an engine, a tree of items — each either presenting a declared item of the form, or holding other items, or standing on its own — and an optional catalogue of translations. Text travels as codes, never as sentences, and the rules that need the form itself (an item exists, a widget is one the engine draws) are reported as `presentation.*` findings.'
      type: object
...
    ProblemError:
      description: "One validation error, mapped 1:1 from the engine's report or a constraint violation."
      required:
        - pointer
        - code
        - message
      properties:
        pointer:
          description: 'RFC 6901 JSON Pointer into the offending document ("" is its root).'
          type: string
        code:
          description: 'Machine-readable code, e.g. `schema.required`, `request.range` or `form.field.duplicate-name`.'
          type: string
        message:
          type: string
        input:
          description: 'The offending value — echoed only when it is a scalar.'
          type:
            - string
            - number
            - boolean
      type: object
      additionalProperties: false
    Problem:
      description: 'RFC 9457 problem details. `type` is a URN `urn:problem:ingot-forms:<slug>`; validation problems carry the `errors` extension member.'
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ProblemError'
      type: object
      additionalProperties: false
    ActivateVersionsRequest:
      required:
        - definition
      properties:
        definition:
          description: 'Which published definition new forms are made of from here on.'
          type: integer
          example: 3
          exclusiveMinimum: 0
        presentation:
          description: 'Which published presentation shows them. Leave it out and this template shows nothing — the pair is stated whole, so an omitted presentation means none rather than the one in use.'
          type:
            - integer
            - 'null'
          example: 7
          default: null
          exclusiveMinimum: 0
      type: object
      additionalProperties: false
...
    CreateFormTemplateRequest:
      required:
        - name
        - definition
      properties:
        name:
          description: 'What to call this template where a person reads it. A label and never an identifier: nothing looks a template up by name, two templates may share one, and it can be changed afterwards without anything a form asks changing with it.'
          type: string
          example: 'Damage report'
          maxLength: 255
        definition:
          description: 'The first definition this template publishes, per the meta-schema this API serves at `GET /api/schemas/definition`. It becomes version 1 and is put in use at once.'
          minProperties: 1
          type: object
          example:
            items:
              -
                type: text
                name: email
                required: true
        presentation:
          description: 'The first presentation, judged against the definition beside it. Optional — a template a system only ever fills in over JSON needs none, and one published later can be added to the other history.'
          type:
            - object
            - 'null'
          example:
            engine: core-html
            items:
              -
                name: email
                widget: text
              -
                widget: confirm
          default: null
      type: object
      additionalProperties: false
    PublishDefinitionRequest:
      required:
        - definition
      properties:
        definition:
          description: "The definition to publish, per `GET /api/schemas/definition`. It is added to this template's definition history with the next number and changes nothing about what forms created now are made of."
          minProperties: 1
          type: object
          example:
            items:
              -
                type: text
                name: email
                required: true
      type: object
      additionalProperties: false
    PublishPresentationRequest:
      required:
        - presentation
      properties:
        presentation:
          description: 'The presentation to publish, per `GET /api/schemas/presentation`. Judged against the definition currently in use; a presentation showing an item that definition does not declare is refused here, with the findings pointing at the item.'
          minProperties: 1
          type: object
          example:
            engine: core-html
            items:
              -
                name: email
                widget: text
              -
                widget: confirm
      type: object
      additionalProperties: false
...
  responses:
    MalformedJson:
      description: 'The request body is not valid JSON, or its media type is missing.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            malformed-json:
              summary: 'Unparsable body'
              value:
                type: 'urn:problem:ingot-forms:malformed-json'
                title: 'Request body is not valid JSON.'
                status: 400
    UnsupportedMediaType:
      description: 'The request body is not `application/json` — no other media type is accepted.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            unsupported-media-type:
              summary: 'A form-encoded or missing content type'
              value:
                type: 'urn:problem:ingot-forms:unsupported-media-type'
                title: 'Only application/json request bodies are accepted.'
                status: 415
...
    FormTemplateNotFound:
      description: 'No form template with this id.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            form-template-not-found:
              summary: 'Unknown id'
              value:
                type: 'urn:problem:ingot-forms:form-template-not-found'
                title: 'Form template not found.'
                status: 404
    TemplateVersionNotFound:
      description: 'This template published no such version. A number is a place in one of its two histories, and an unknown one is a missing document rather than a bad request.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            template-version-not-found:
              summary: 'A number this template never handed out'
              value:
                type: 'urn:problem:ingot-forms:template-version-not-found'
                title: 'The form template has no such version.'
                status: 404
    FormGone:
      description: 'The form has expired; its data is scheduled for physical deletion.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            form-gone:
              summary: 'Past its expire date'
              value:
                type: 'urn:problem:ingot-forms:form-gone'
                title: 'Form has expired.'
                status: 410